adding docker support, adding support to install unixodbc and redshift drivers into docker, addweb ui support for c3 and d3, adding a large number of unit test

+++ reporting-test-client/IrrigationReportingWebApi/DataContracts/IBearingMinuiteConverter.cs
@@ -2,7 +2,7 @@
 {
 	public interface IBearingMinuiteConverter
 	{
-		double ToDegrees(int? bearingMinutes);
+		decimal ToDegrees(int? bearingMinutes);
 		int ToBearingMinutes(double bearing);
 	}
 }
