+++ reporting-test-client/IrrigationReportingWebApi/DataContracts/IrrigationEvent.cs
@@ -5,7 +5,7 @@ namespace Trimble.Ag.IrrigationReporting.DataContracts
 	public class IrrigationEvent
 	{
 		public Int64 JournalId { get; set; }
-		public double Bearing { get; set; }
+		public decimal Bearing { get; set; }
 		public int? RotationId { get; set; }
 		public double? Velocity { get; set; }
 		public bool? Pump { get; set; }
